micropython/requirements.txt
gunicorn
tensorflow
ultralytics
torch
pillow
torchvision
flask
google-cloud-storage
gevent
uvicorn